WIP: Hack to test companion pair verify
This commit introduces a lot of hacks to test the verification part of
encryption for the companion protocol. Verification seems to work with
credentials that have been obtained prior with pairing. Data encryption
has not yet been tested as I don't know what payload to send. I intend
to implement support in the proxy to figure that out.

A lot of codes is just copied from MRP, so I need to spend some time to
consolidate and share code when I have some time.